RUST-2245: Separate windows and non-windows implementations
